Add Float variable to InputContext and defined getters with templates
The current implementation forces the developer to use one of the available types to store data inside the InputContext. When getting the value by a type that it is not available, the getter crashes (as intended).
However, this solution is not scalable and each InputSource might want to add multiple values to the  InputContext (e.g. when using axis, one might want to save rawDelta, normalizedDelta, axisPositions, etc...) In the future, the input manager should allow for this features.
update get value