feat(ecs): implement proper Archetypes
This commit is huge and definitely should have been split into many
smaller commits, but there were so many interdependencies between
stuff that needed to be changed that splitting it would require way
longer than we have time for.

It essentially changes the entire ECS structure to be archetype based,
which paves the way for implementing entity relations.