refactor(reflection): no longer fallback to anonymous types automatically
This change was necessary due to the previous system being highly error prone.
Before, if reflect<T> was called for a type T without defined reflection included, a placeholder definition would be used which returned an anonymous type.
Sadly, due to the C++ One Definition Rule, this meant that other source files using reflect<T>, could return an anonymous type, even if the type's reflection definition was included!

TL;DR, previously a simple mistake such as forgetting to include an external reflection header could lead to a random reflection failure in the core of the library, such as a missing trait error.
Now, the users must define reflection for all types, as there is no longer a fallback. To simplify this for the previous 'anonymous types', a new CUBOS_ANONYMOUS_REFLECT macro was added.