fix(ecs): remove duplicate query matches on self-relations
Previously the behavior queries had with symmetric relations on entities was not consistent.
For example, if we had a symmetric relation R, and an entity 0, with a relation from 0 to 0, then, when querying over relation R, we would get (0, 0) twice!

This only happened when the query was asymmetric, i.e, when the requirements on both targets were different. If they were the same, we would only get (0, 0).

This lead to some inconsistencies between the tests and the query code.
I also think it was a bit confusing, so I changed it to always return only one match in these cases.